updating the PlantDetails, TaskHistory and HarvestTracker
