+++ frontend/src/pages/GardenDetails/GardenDetails.css
@@ -11,16 +11,17 @@
 
 .plant-picture {
   object-fit: cover;
-  width: 290px;
-  height: 395px; /* Adjust the height as per your requirement */
+  width: 300px;
+  height: 400px; /* Adjust the height as per your requirement */
 
 }
 
 .bottom-container {
   display: flex;
-  flex-wrap: wrap; /* Allow cards to wrap to the next line */
-  justify-content: left; /* Center the cards horizontally */
+  flex-wrap: wrap; 
+  justify-content: left; 
   margin-top: 40px;
+  justify-content: space-evenly;
 }
 
 .description {
@@ -28,6 +29,7 @@
   padding: 10px;
   margin-bottom: 10px;
   justify-content: flex-start;
+  margin: 20px;
 }
 
 .add-plant {
@@ -35,6 +37,7 @@
   padding: 10px;
   margin-bottom: 10px;
   justify-content: center;
+  margin: 20px;
 }
 
 .users {
@@ -42,6 +45,7 @@
   padding: 10px;
   margin-bottom: 10px;
   justify-content: flex-end;
+  margin: 20px;
 }
